% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_helpers.R
\name{all_after}
\alias{all_after}
\title{Shortcuts for stringr regex. Identifies substring occuring before or after character.
Note: these are technically all_after_last() and all_before_first() occurances.
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}}
\usage{
all_after(character)
}
\value{
A vector.
}
\description{
Shortcuts for stringr regex. Identifies substring occuring before or after character.
Note: these are technically all_after_last() and all_before_first() occurances.
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
str_extract('string_test',all_after("_"))
str_remove('string_test',all_after("_"))

str_extract('string/test',all_after("/"))
str_extract('string/test',all_before("/"))
str_extract('string_test',all_before("_"))
str_remove('string_test',all_before("_"))
str_extract('string_test_two', all_after("_"))
str_remove('string_test_two',  all_after("_"))
str_extract('string_test_two', all_before("_"))
str_remove('string_test_two',  all_before("_"))
}
