% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggShortCuts.R
\docType{data}
\name{bottom_legend}
\alias{bottom_legend}
\title{Custom legend & axis shortcuts}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
bottom_legend
}
\description{
Custom legend & axis shortcuts
}
\keyword{datasets}
