% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gg_colours.R
\name{brewer_GG_col}
\alias{brewer_GG_col}
\title{Select R color brewer palette of type my_palette with nCols n colours}
\usage{
brewer_GG_col(my_palette = "Paired", nCols)
}
\arguments{
\item{my_palette}{An RColourBrewer palette name (see RColorBrewer::display.brewer.all())}

\item{nCols}{Number of colours to select from palette. These will be evenly spaced across palette.}
}
\description{
Select R color brewer palette of type my_palette with nCols n colours
}
