% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_GG_col.R
\name{default_GG_col}
\alias{default_GG_col}
\title{Select ggplot default colour palette of n. colours}
\usage{
default_GG_col(x)
}
\arguments{
\item{x}{N colours to select}
}
\description{
Select ggplot default colour palette of n. colours
}
