% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_grey.R
\name{default_grey}
\alias{default_grey}
\alias{default_gray}
\title{default_grey}
\usage{
default_grey()

default_gray()
}
\value{
A string
}
\description{
Quick access to ggplot default grey/gray (hex #595959).
}
\section{Functions}{
\itemize{
\item \code{default_gray()}: Returns #595959

}}
\examples{
scales::show_col(default_gray())
scales::show_col(default_grey())
}
