% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxdodge.R
\name{geom_boxdodge}
\alias{geom_boxdodge}
\title{Create boxplots with multiple series on x axis, overlaid with datapoints, displaying exact y values and jittered x values.}
\usage{
geom_boxdodge(..., point_size = 0.25, box_alpha = 0.5)
}
\arguments{
\item{point_size}{size of points; default is 0.25}

\item{box_alpha}{alpha for boxplot background ; default is 0.5}
}
\value{
A ggplot.
}
\description{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
require(ggplot2)
mpg |> ggplot2::ggplot(aes(x=class, y=cty, col=factor(year))) +
 geom_boxdodge(point_size = 3,box_alpha = 0.75)
}
