% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_errorbar_auto.R
\name{geom_error_sd}
\alias{geom_error_sd}
\alias{geom_error_se}
\title{ggplot error bars}
\usage{
geom_error_sd(...)

geom_error_se(...)
}
\value{
A grob (graphical object).
}
\description{
Draw error bars as standard error or standard deviation. cf https://stackoverflow.com/a/19299034/20620467.
Adapted from \href{https://rpkgs.datanovia.com/ggpubr/}{ggpubr}.
}
\section{Functions}{
\itemize{
\item \code{geom_error_sd()}: Draw standard deviation bars

\item \code{geom_error_se()}: Draw standard error bars

}}
