% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head_tail}
\alias{head_tail}
\title{Show head and tail simultaneously}
\usage{
head_tail(data, n_slice = 6)
}
\arguments{
\item{data}{The data frame to slice}

\item{n_slice}{How many rows of the head and tail you want}
}
\value{
A data frame of the first and last n_slice rows
}
\description{
Show head and tail simultaneously
}
\details{
This will work on matrices and also not be limited to default tibble display.
}
\examples{

library(tidyext)
as.matrix(mtcars) \%>\%
head_tail(6)

}
