% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle_col.R
\name{plot_cycle_col}
\alias{plot_cycle_col}
\title{Recycle color scales over multiple scatter plots. Use cases include PCA / MDS plots with multiple factors.}
\usage{
plot_cycle_col(df, X, Y, myLabel, colour_vec)
}
\arguments{
\item{df}{data frame}

\item{X}{x axis mapping}

\item{Y}{y axis mapping}

\item{myLabel}{TBC: labelling functionality}

\item{colour_vec}{vector of values to colour points by.}
}
\value{

}
\description{
Recycle color scales over multiple scatter plots. Use cases include PCA / MDS plots with multiple factors.
}
\examples{
my_df <- mpg \%>\% mutate(year=factor(year), cyl=factor(cyl))
my_features <- c('year','drv','cyl','manufacturer')

plot_cycle_col(df = my_df,  X='cty',Y='hwy', myLabel = 'manufacturer', colour_vec = my_features)
}
