% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle_cols.R
\name{plot_cycle_cols}
\alias{plot_cycle_cols}
\title{Recycle color scales over multiple scatter plots}
\usage{
plot_cycle_cols(df, X, Y, myLabel, colour_vec)
}
\arguments{
\item{df}{data frame}

\item{X}{x axis mapping}

\item{Y}{y axis mapping}

\item{myLabel}{TBC: labelling functionality}

\item{colour_vec}{vector of values to colour points by.}
}
\description{
Use cases include PCA / MDS plots with multiple factors
}
\details{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
my_df <- mpg \%>\% mutate(year=factor(year), cyl=factor(cyl))
my_features <- c('year','drv','cyl','manufacturer')

plot_cycle_cols(df = my_df,  X='cty',Y='hwy', myLabel = 'manufacturer', colour_vec = my_features)
}
