% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{Print all data frame rows to console}
\usage{
print_all(x, n = nrow(x))
}
\value{
A data frame.
}
\description{
Print all data frame rows to console
}
\examples{
mpg |> print_all()
mpg |> print_all(n=50)
}
