% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_anti.R
\name{recip_anti}
\alias{recip_anti}
\title{Print values in vector 1 that are not in vector 2.}
\usage{
recip_anti(x_vec, y_vec)
}
\arguments{
\item{x_vec}{vector of elements of any type}

\item{y_vec}{vector of elements of any type}
}
\value{
A tibble.
}
\description{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
