% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_order.R
\name{recip_order}
\alias{recip_order}
\title{Report whether the order of two vectors is identical.}
\usage{
recip_order(x_vec, y_vec)
}
\arguments{
\item{x_vec}{vector of elements of any type.}

\item{y_vec}{vector of elements of any type.}
}
\value{
A message.
}
\description{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
recip_order(mtcars$mpg,mtcars$mpg)
recip_order(mtcars$mpg,sort(mtcars$mpg))
recip_order(mtcars$mpg,mtcars$mpg[1:5])
}
