% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd}
\alias{rnd}
\title{Sensible rounding for printing.}
\usage{
rnd(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Digits to round to. Default is 3.}
}
\value{
A character vector suitable for printed output.
}
\description{
Sensible rounding for printing.
}
\details{
This is the rounding many are wanting for printing out results. For
  example, there will always be \code{digits} number of decimal places, and
  trailing zeros will be kept.
}
\examples{
library(tidyext)
set.seed(123)
rnd(rnorm(10))
rnd(1)
}
