% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_pct.R
\name{sort_pct}
\alias{sort_pct}
\title{Count any number of columns, print percentage and sort descending}
\usage{
sort_pct(.df, ...)
}
\description{
Count any number of columns, print percentage and sort descending
}
