% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tidy.R
\name{summary_tidy}
\alias{summary_tidy}
\title{Tibble output for baseR summary() function.}
\usage{
summary_tidy(tbl, data_type = "cont")
}
\value{
A tibble.
}
\description{
Default is to return summary stats for all columns containing continuous data.
Option to specify categorical data.
}
\examples{
require(ggplot2)
mpg |> summary_tidy(data_type = 'cont')
mpg |> summary_tidy(data_type = 'categ')

}
