% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyext.R
\docType{package}
\name{tidyext}
\alias{tidyext}
\alias{tidyext-package}
\title{tidyext: Common data processing in the tidyverse.}
\description{
This package is a collection of functions that do the things I commonly need
to do with data. I work with data for myself and others everyday, and use
these functions quite often.  The goal here is to not have any dependencies beyond the tidyverse.
}
\details{
To learn more about tidyext, you can browse the vignettes whenever they actually
get created some day: \code{browseVignettes(package = "tidyext")}
}
