% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_rows.R
\name{unite_rows}
\alias{unite_rows}
\title{Convert vector elements to a delimited string.}
\usage{
unite_rows(.data, input_colname, output_colname, group_var, delim = ";")
}
\arguments{
\item{.data}{A tibble}

\item{input_colname}{The column name to collapse into rows}

\item{output_colname}{The new column name to contain collapsed data}

\item{group_var}{Any number of grouping columns}

\item{delim}{delimeter for collapsed data}
}
\value{
If no groups are specified, returns a vector; else returns an ungrouped dataframe.
}
\description{
This function performs the converse to separate_rows() / \href{https://tidyr.tidyverse.org/reference/separate_longer_delim.html}{separate_longer_delim()} functions. Adapted from https://stackoverflow.com/a/56810542.
Note this function allows non-standard evaluation. Quotes for column names and groups are optional.
}
\details{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
mpg \%>\% unite_rows(displ, displ_str, class)
mpg \%>\% unite_rows(displ, displ_str, c(class, cyl), delim=',')
}
