% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_legend.R
\name{bottom_legend}
\alias{bottom_legend}
\title{Bottom legend}
\usage{
bottom_legend()
}
\description{
Moves ggplot2 legend to bottom of plotting space.
}
\details{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
mpg \%>\% ggplot(aes(x=cty,y=hwy)) + geom_point(aes(col=drv)) + bottom_legend()
}
