% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxjitter.R
\name{geom_boxjitter}
\alias{geom_boxjitter}
\title{Create boxplots overlaid with datapoints, displaying exact y values and jittered x values}
\usage{
geom_boxjitter(
  ...,
  point_size = 0.5,
  point_col = "black",
  box_alpha = 1,
  jitterheight = 0,
  jitterwidth = 0.25
)
}
\arguments{
\item{point_size}{size of points; default is 0.5}

\item{point_col}{point colour; default is black}

\item{box_alpha}{alpha for boxplot background ; default is 1}

\item{jitterheight}{jitter height; default is 0}

\item{jitterwidth}{jitter width; default is 0.25}
}
\value{
A ggplot.
}
\description{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
require(ggplot2)
mpg |> ggplot2::ggplot(aes(x=class, y=cty)) +
 geom_boxjitter(notch = FALSE, point_size = 2, point_col='dodger blue')
}
