% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_errorbar_auto.R
\name{geom_error_sd}
\alias{geom_error_sd}
\alias{geom_error_se}
\title{ggplot error bars}
\usage{
geom_error_sd(...)

geom_error_se(...)
}
\value{
A grob (graphical object).
}
\description{
Draw error bars as standard error or standard deviation. cf https://stackoverflow.com/a/19299034/20620467.
Adapted from \href{https://rpkgs.datanovia.com/ggpubr/}{ggpubr}.
}
\section{Functions}{
\itemize{
\item \code{geom_error_sd()}: Draw standard deviation bars

\item \code{geom_error_se()}: Draw standard error bars

}}
\examples{
ggplot2::ggplot(mtcars, aes(x=cyl,y= qsec)) +
  geom_jitter(alpha=0.25,width=0.25) +
  geom_error_se(col='red', lwd=1,width=0.4) +
  geom_error_sd(col='blue',lwd=1,width=0.6) +
  geom_point(stat='summary',fun=mean,col='purple') +
  ggtitle('error bars',
  subtitle = '±1 standard deviation in blue;\n±1 standard error of the mean in red')

}
