% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_table.R
\name{recip_table}
\alias{recip_table}
\title{Reciprocal table function: quantifies the intersect of unique occurrences between two vectors}
\usage{
recip_table(x_vec, y_vec)
}
\arguments{
\item{x_vec}{vector of values}

\item{y_vec}{vector of values}
}
\value{
A tibble.
}
\description{
Reciprocal table function: quantifies the intersect of unique occurrences between two vectors
}
\examples{
recip_table(mtcars$mpg,mtcars$mpg[1:5])
recip_table(mtcars$mpg[1:5], mtcars$mpg)
recip_table(mtcars$mpg, mtcars$mpg)
recip_table(mtcars$mpg, mtcars$cyl)
}
