% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyExt-package.R
\docType{package}
\name{tidyExt}
\alias{tidyExt}
\title{tidyExt}
\description{
tidyExt: a host of helper functions for the tidyverse, designed to
overcome some common pain-points in plotting, data summarization and environment management.
}
