% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyExt-package.R
\docType{package}
\name{tidyExt}
\alias{tidyExt}
\title{tidyExt}
\description{
tidyExt: a host of helper functions for the tidyverse, designed to
overcome some common pain-points in plotting, data summarization and environment management.
}
\details{
\href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}

\strong{Functions for managing source code and global environment}
\itemize{
\code{\link{printScriptDir }}

\code{\link{printScriptDir}}

\code{\link[ggpubr]{stat_cor}}

\code{\link{fix_tidyverse_conflicts}}
}

\strong{Data view helpers}
\itemize{
\code{\link{bighead}}

\code{\link{print_all}}
}

\strong{Statistics helpers}
\itemize{
\code{\link{scale_this}}

\code{\link{sort_pct}}
}

\strong{Faceted scatterplot colour recycling}
\itemize{
\code{\link{plot_cycle_cols}}
}

\strong{GGplot2 colour helpers}
\itemize{
\code{\link{default_GG_col}}

\code{\link{brewer_GG_col}}
}

\strong{GGplot2 shortcuts}
\itemize{
\code{\link{geom_boxjitter}}

\code{\link{geom_boxdodge}}

\code{\link{geom_smooth_lm}}

\code{\link{x_angle}}

\code{\link{no_legend}}

\code{\link{bottom_legend()}}
}
}
